+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+### Fixed
+- fix TypeError from being thrown if the position of definition, highlight range, or rename calculations are from outside the document ([#132](https://github.com/rcjsuen/dockerfile-language-service/issues/132))
+
 ## [0.15.0] - 2025-04-10
 ### Added
 - support preparing renames for here-documents ([#129](https://github.com/rcjsuen/dockerfile-language-service/issues/129))